+++ src/main/java/com/ngtesting/platform/service/impl/OrgRoleUserServiceImpl.java
@@ -36,7 +36,7 @@ public class OrgRoleUserServiceImpl extends BaseServiceImpl implements OrgRoleUs
         	vo.setSelected(false);
         	vo.setSelecting(false);
         	for (TestUser po2 : orgRoleUsers) {
-        		if (po1.getId() == po2.getId()) {
+        		if (po1.getId().longValue() == po2.getId().longValue()) {
             		vo.setSelected(true);
             		vo.setSelecting(true);
             	}
